+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* added ReadFromTextFileCredentials
 * added lambda on exception in retryTransaction
 
 ## 1.10.0